requirements.txt
numpy==2.2.5
matplotlib==3.10.1
scikit-learn==1.6.1
pandas==2.2.3
tensorflow==2.15.0
tf-keras==2.15.0
tensorflow-addons==0.23.0
tifffile
imageio
albumentations
torchinfo
tqdm
